sanheyi/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
pillow-avif-plugin==1.4.3
reportlab==4.0.9
resvg_py==0.2.0
python-pptx==0.6.23
python-docx==1.1.0
openpyxl==3.1.2
ebooklib==0.18
odfpy==1.4.1
pytoshop==1.2.1
numpy
pyinstaller>=6.3.0
httpx==0.26.0
